Random movement now can move in any direction
Random movement can now move in more directions other than the cardinal and ordinal directions. Also cleaned up EnemyMovement.

+++ Assets/Scripts/.vshistory/EnemyMovement.cs/2024-07-31_13_42_01_722.cs
@@ -20,6 +20,8 @@ public class EnemyMovement : MonoBehaviour
 
     private float countdown; // Used for counting down the timer
 
+    //private static List<Vector3> directions; // List of vectors for direction
+
     // Start is called before the first frame update
     void Start()
     {
@@ -42,7 +44,7 @@ public class EnemyMovement : MonoBehaviour
     private void RandomMovement()
     {
         // Move if far away enough from the player
-        float distance = Vector3.Distance(player.position, transform.position);
+        int distance = Vector3.Distance(player.position, transform.position);
         if(distance >= range)
         {
             // Switch direction after countdown
